+++ test/influxdb/write_api_batching_test.rb
@@ -25,7 +25,7 @@ class WriteApiBatchingTest < MiniTest::Test
     WebMock.disable_net_connect!
 
     @write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
-                                                 batch_size: 2, flush_interval: 5_000)
+                                                 batch_size: 2, flush_interval: 5_000, retry_interval: 2_000)
     @client = InfluxDB2::Client.new('http://localhost:9999',
                                     'my-token',
                                     bucket: 'my-bucket',
@@ -42,6 +42,29 @@ class WriteApiBatchingTest < MiniTest::Test
     assert_equal true, @write_client.closed
 
     WebMock.reset!
+  end
+
+  def test_batch_configuration
+    error = assert_raises ArgumentError do
+      InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING, batch_size: 0)
+    end
+    assert_equal "The 'batch_size' should be positive or zero, but is: 0", error.message
+
+    error = assert_raises ArgumentError do
+      InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING, flush_interval: -10)
+    end
+    assert_equal "The 'flush_interval' should be positive or zero, but is: -10", error.message
+
+    error = assert_raises ArgumentError do
+      InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING, retry_interval: 0)
+    end
+    assert_equal "The 'retry_interval' should be positive or zero, but is: 0", error.message
+
+    InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING, jitter_interval: 0)
+    error = assert_raises ArgumentError do
+      InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING, jitter_interval: -10)
+    end
+    assert_equal "The 'jitter_interval' should be positive number, but is: -10", error.message
   end
 
   def test_batch_size
@@ -162,5 +185,108 @@ class WriteApiBatchingTest < MiniTest::Test
                      times: 1, body: "h2o_feet,location=coyote_creek level\\ water_level=1.0 1\n" \
                      "h2o_feet,location=coyote_creek level\\ water_level=2.0 2\n" \
                      'h2o_feet,location=coyote_creek level\\ water_level=3.0 3')
+  end
+
+  def test_retry_interval_by_config
+    error_body = '{"code":"temporarily unavailable","message":"Token is temporarily over quota. '\
+                 'The Retry-After header describes when to try the write again."}'
+
+    stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_return(status: 429, headers: { 'X-Platform-Error-Code' => 'temporarily unavailable' }, body: error_body).then
+      .to_return(status: 204)
+
+    request = "h2o_feet,location=coyote_creek water_level=1.0 1\n" \
+               'h2o_feet,location=coyote_creek water_level=2.0 2'
+
+    @write_client.write(data: ['h2o_feet,location=coyote_creek water_level=1.0 1',
+                               'h2o_feet,location=coyote_creek water_level=2.0 2'])
+
+    sleep(0.5)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
+
+    sleep(1)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
+
+    sleep(1)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 2, body: request)
+  end
+
+  def test_retry_interval_by_header
+    error_body = '{"code":"temporarily unavailable","message":"Server is temporarily unavailable to accept writes. '\
+                 'The Retry-After header describes when to try the write again."}'
+
+    stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_return(status: 503, headers: { 'X-Platform-Error-Code' => 'temporarily unavailable', 'Retry-After' => '3' },
+                 body: error_body).then
+      .to_return(status: 204)
+
+    request = "h2o_feet,location=coyote_creek water_level=1.0 1\n" \
+               'h2o_feet,location=coyote_creek water_level=2.0 2'
+
+    @write_client.write(data: ['h2o_feet,location=coyote_creek water_level=1.0 1',
+                               InfluxDB2::Point.new(name: 'h2o_feet')
+                                   .add_tag('location', 'coyote_creek')
+                                   .add_field('water_level', 2.0)
+                                   .time(2, InfluxDB2::WritePrecision::NANOSECOND)])
+
+    sleep(0.5)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
+
+    sleep(1)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
+
+    sleep(1)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
+
+    sleep(1)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 2, body: request)
+  end
+
+  def test_jitter_interval
+    @client.close!
+
+    @client = InfluxDB2::Client.new('http://localhost:9999',
+                                    'my-token',
+                                    bucket: 'my-bucket',
+                                    org: 'my-org',
+                                    precision: InfluxDB2::WritePrecision::NANOSECOND,
+                                    use_ssl: false)
+
+    @write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                 batch_size: 2, flush_interval: 5_000, jitter_interval: 2_000)
+    @write_client = @client.create_write_api(write_options: @write_options)
+
+    stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_return(status: 204)
+
+    request = "h2o_feet,location=coyote_creek water_level=1.0 1\n" \
+               'h2o_feet,location=coyote_creek water_level=2.0 2'
+
+    @write_client.write(data: ['h2o_feet,location=coyote_creek water_level=1.0 1',
+                               'h2o_feet,location=coyote_creek water_level=2.0 2'])
+
+    sleep(0.1)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 0, body: request)
+
+    sleep(2)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
   end
 end