+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 3. [#19](https://github.com/influxdata/influxdb-client-ruby/issues/19): Added WriteOptions and possibility to batch write
 
 ### Bugs
-1. [](https://github.com/influxdata/influxdb-client-ruby/issues/): Fixed batch write
+1. [#22](https://github.com/influxdata/influxdb-client-ruby/pull/22): Fixed batch write
  
 ## 1.0.0.beta [2020-01-17]