+++ CHANGELOG.md
@@ -1,14 +1,14 @@
 ## 1.2.0 [unreleased]
 
+### Bugs
+1. [#22](https://github.com/influxdata/influxdb-client-ruby/pull/22): Fixed batch write
+
 ## 1.1.0 [2020-02-14]
 
 ### Features
 1. [#14](https://github.com/influxdata/influxdb-client-ruby/issues/14): Added QueryApi
 2. [#17](https://github.com/influxdata/influxdb-client-ruby/issues/17): Added possibility to stream query result
 3. [#19](https://github.com/influxdata/influxdb-client-ruby/issues/19): Added WriteOptions and possibility to batch write
-
-### Bugs
-1. [#22](https://github.com/influxdata/influxdb-client-ruby/pull/22): Fixed batch write
  
 ## 1.0.0.beta [2020-01-17]